requirements.txt
psutil
GPUtil
colorama
whois
dnspython
mysql-connector-python
rich
google-generativeai
prompt-toolkit
selenium
webdriver-manager
faker